feat(cli,config): wire node presentation discipline end-to-end + proptest (DEV-C2)
- multiview-cli: the node runner consumes `timing.link_offset_ms` —
  `Pipeline::set_node_presentation` enables the pull-side frame chooser for the
  run's display heads (node-only; a plain `multiview run` stays undisciplined
  latest-wins). `start_display_sinks` builds one `PresentationPlan` per head
  (shared outbound epoch + link offset + a real CLOCK_MONOTONIC/REALTIME
  `RealtimePresentationClock`); `state_of` publishes each canvas with its
  output-PTS via `publish_at` so the chooser can compute deadlines.
  Epoch-source boundary (stated honestly): the epoch is the run's LOCAL outbound
  presentation epoch — real and self-contained for a single node; a future
  controller WS-client (DEV-B6) fills the SAME `SharedEpoch` seam, nothing else
  changes. A lost feed free-runs on the last epoch (inv #1/#10).
- multiview-config: NodeTiming.link_offset_ms docs now describe the consumed
  behaviour (was 'recorded-not-consumed until DEV-C2').
- display_present_props: proptest for the pure logic (guardrails) — choice index
  always in range, present-is-nearest argmin, repeat-early iff >½-period-ahead,
  degenerate-period newest-wins, prediction is the next on-grid instant strictly
  after now within one period, queue bounded + newest-wins + drop-oldest,
  deadlines == wall_at(pts)+link. 8/8 green; no regressions generated.

display-kms clippy clean (cli+output, -D warnings); cli default build green.

+++ crates/multiview-output/src/display/present.rs
@@ -363,9 +363,12 @@ impl RealtimePresentationClock {
     /// One `clock_gettime` reading as integer nanoseconds.
     fn read_ns(clock: rustix::time::ClockId) -> i64 {
         let ts = rustix::time::clock_gettime(clock);
+        // `display-kms` is Linux-only, where `rustix::time::Timespec` carries
+        // both fields as `i64` (`tv_nsec` is always `0..1e9`), so the seconds
+        // term cannot realistically overflow; `saturating_*` is belt-and-braces.
         ts.tv_sec
             .saturating_mul(1_000_000_000)
-            .saturating_add(i64::from(ts.tv_nsec))
+            .saturating_add(ts.tv_nsec)
     }
 }
